optimization of the joins if you have a query like SELECT x FROM Type t JOIN t.a.b.c x
OQL joins also the table of x (t.a.b.c) . But if no field of x is selected
(only the primary key), the primary key is available in the table of t.a.b,
hence the "last" table doesn't need to be joined.

I learned this optimization by looking at Hibernate-generated SQL :)

Also in the process I found that there are quite many unused labels (they
appear in FROM but are never used in WHERE, SELECT, etc). Those are not joined
any longer. A warning is issued.